fix: Restructure components in a more logical way
