+++ src/components/PollHeader.js
